+++ src/banner/propPane/index.ts
@@ -3,3 +3,8 @@
 // export { prepSourceColumns } from './prepSourceColumns';
 
 // export * from '';
+
+export { FPSOptionsGroupBasic, } from './FPSOptionsGroupBasic';
+export { FPSBanner4BasicGroup,  } from './FPSBanner4BasicGroup';
+export { FPSBanner3NavGroup, } from './FPSBanner3NavGroup';
+export { FPSBanner3ThemeGroup } from './FPSBanner3ThemeGroup';